+++ src/BlockHorizons/BlockSniper/commands/BrushCommand.php
@@ -12,9 +12,10 @@ use BlockHorizons\BlockSniper\ui\windows\BrushMenuWindow;
 use BlockHorizons\BlockSniper\ui\windows\MainMenuWindow;
 use pocketmine\command\CommandSender;
 use pocketmine\item\Item;
-use pocketmine\nbt\tag\IntTag;
+use pocketmine\nbt\tag\StringTag;
 use pocketmine\Player;
 use pocketmine\utils\TextFormat as TF;
+use pocketmine\utils\UUID;
 
 class BrushCommand extends BaseCommand{
 
@@ -35,28 +36,29 @@ class BrushCommand extends BaseCommand{
 					$sender->sendMessage($this->getWarning() . Translation::get(Translation::COMMANDS_BRUSH_NEED_ITEM));
 					return;
 				}
-				$b = new Brush($sender->getName());
-
-				$brushIndex = BrushListener::$brushIndex++;
-				BrushListener::$brushItems[$brushIndex] = $b;
-				$sender->sendForm(new BrushMenuWindow($this->loader, $sender, $b, true));
-
-				$item->getNamedTag()->setInt(BrushListener::KEY_BRUSH_ID, $brushIndex);
-				$sender->getInventory()->setItemInHand($item);
-				break;
-			case $args[0] === "unbind":
-				if(!$item->getNamedTag()->hasTag(BrushListener::KEY_BRUSH_ID, IntTag::class)){
-					$sender->sendMessage($this->getWarning() . Translation::get(Translation::COMMANDS_BRUSH_NOT_BOUND));
-					return;
-				}
 				$brush = $this->loader->config->brushItem->parse();
 				if($item->getId() === $brush->getId() && $item->getDamage() === $brush->getDamage()){
 					$sender->sendMessage($this->getWarning() . Translation::get(Translation::COMMANDS_BRUSH_BIND_BRUSH_ITEM));
 					return;
 				}
-				$index = $item->getNamedTag()->getInt(BrushListener::KEY_BRUSH_ID);
-				$item->getNamedTag()->removeTag(BrushListener::KEY_BRUSH_ID);
-				unset(BrushListener::$brushItems[$index]);
+				$b = new Brush($sender->getName());
+
+				$uuid = UUID::fromRandom()->toString();
+				BrushListener::$brushItems[$uuid] = $b;
+				$sender->sendForm(new BrushMenuWindow($this->loader, $sender, $b, true));
+
+				$item->getNamedTag()->setString(BrushListener::KEY_BRUSH_UUID, $uuid);
+				$sender->getInventory()->setItemInHand($item);
+				break;
+			case $args[0] === "unbind":
+				if(!$item->getNamedTag()->hasTag(BrushListener::KEY_BRUSH_UUID, StringTag::class)){
+					$sender->sendMessage($this->getWarning() . Translation::get(Translation::COMMANDS_BRUSH_NOT_BOUND));
+					return;
+				}
+				$uuid = $item->getNamedTag()->getString(BrushListener::KEY_BRUSH_UUID);
+				$item->getNamedTag()->removeTag(BrushListener::KEY_BRUSH_UUID);
+				unset(BrushListener::$brushItems[$uuid]);
+
 				$sender->sendMessage(TF::GREEN . Translation::get(Translation::COMMANDS_BRUSH_CLEAR_SUCCESS));
 		}
 	}